+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - **BC break**: Split subject-line preview SDK out to
   [`phlib/litmus-subjectline-preview`](https://github.com/phlib/litmus-subjectline-preview)
   as it doesn't use the reseller API.
+- **BC break**: Remove PageTest classes as this isn't in the Litmus API.
 
 ## [2.0.1] - 2021-03-30